Tried incorporating a ConstVectorKeys for immutable leaf collections alongside a mutable VectorKeys. Ended up with generics hell. I mean _hell_. I dare you to look at ConstMapCIA, or CIAm, after expanding ExactConstMapCIA. This doesn't even begin to mention what Tree would've been. For kicks, here's the full parameter list:
class Tree<+Tv, Tx as arraykey, TCollection as \ConstMap<Tx, Tv>, TKeyUnderlying as ConstVectorKeys, TKeyWrapperKeyWrapper as \ConstVector<Tx>, TKeyWrapper as \ConstVectorIA<Tx, TKeyWrapperKeyWrapper, TUnderlying>>

*shudders*

One of those six parameters is the shortcutted key iterator for the Vector holding the keys for the map. Come to think of it, three of them are devoted to holding the map keys, making sure it's immutable when the class is immutable. It's a valiant cause -- to ensure references to the keys container in a mutable context isn't allowed to modify anything. However, giving it up would make everything _much_ more readable.

We could do both would be possible with union KeyedIterable and ConstIndexAccess. We could also merge the KeyedContainerWrapper and Iterable[Const]IndexAccess trees together, eliminating the need for the {container}[C]IA classes altogether. While we wait, though, let's try and relax the mutability requirements just a bit.