Repaired blocking loop in `TotalAwaitable`
`TotalAwaitable`, as it's used in `StreamFactory`, does far more than just represent the total execution time of the application (or sub-application). It has to kickstart all of the `Awaitable`s that are presented to it. AllAwaitWaitHandle then must appear so that they are all kicked off simultaneously.

Another big fix is with `add` behaviour, which would suffer the same problem if nothing changed there. The batch of `Awaitable` that is awaited initially would not be updated instantly with new `Awaitable`s with the naive implementation, so the underlying handle must actually be halted and restarted at every add operation.