Repaired `Haltable::soft_halt` behaviour by implementing idempotency.
Note: I've noticed that the LaggingStreams test has regressed ever since the `TotalAwaitable` blocking issue was fixed. Some values emitted from substreams are repeated (usually the first in the collection that is merged), while some are omitted. I think it's a problem arising from resubscribing to the `Producer::next()` awaitable so many times, but it requires much closer inspection.