Fixing lagging collection regression
When producers are waiting on the cutting edge, they _always_ need to shift their lag LL or else they will emit values twice if they are not the first to resolve. This is because the producer to resolve first will add this cutting edge to the shared lag list, and the next iteration, this slower producer will pull this value it's already emitted from the list.