Fixing more regressions for lag: repaired LinkedList start and empty conditions with ugly logic
The previous implementation of LinkedList allowed the head to take on NULL when it surpassed the tail from shifting. However, the desired behaviour is actually that the head preserves the reference to the tail element instead of overstepping into NULL, and to implement `LinkedList::empty` with a flag and some accessory logic instead. This is so that, when elements are tacked on after initially hitting the end, the LinkedList can resume.