Repair error propagation and race condition
- **Error propagation:** While transitioning TotalWaitHandle from
LinkedLists to Vector, I thought it would be more performant to use an
AwaitAllWaitHandle instead of HH\Asio\v, since I didn't care about the
results of awaiting. However, it turns out that AwaitAllWaitHandle
doesn't propagate errors, so I would have to look through each result
anyways to oust the error anyways. As a result, HH\Asio\v is used
instead.
- **Race condition:** In Haltable, there was a glaring race condition
involving the void-, null- and T-Awaitables, such that the `await` call
in the T-Awaitable wasn't guaranteed to resolve before the void-. All
of these have been done away with and replaced by a single Awaitable,
which has the added benefit of reducing the stack size (previously
the Haltable calls took three entries in the stack). Interestingly,
this never manifested in 3.18.0-dev, but did in 3.16.0.