Make minor changes to README, require PHPUnit
The section on `HH\Asio\later` in the README was misleading. `later`
does not ensure total ordering -- only the priority of the scope that
it's awaited. In some cases, this is checked continuously to check
for an eventual behavior (e.g. ConditionWaitHandle::notify), but most
of the time, ordering is controlled by composition of Awaitables.

Unit tests are starting to roll in.