Splitting Producer- and Observable-based impl.
While, when it comes down to it, the Producer-based and Observable-
based implementations of ReactiveX are functionally equivalent and
even similar in form, the Producer-based implementation meshes
more nicely with core Hack constructs. I was occasionally frustrated
by the incompatibility of `Stream`s and `Awaitable`, which I worried
would impose much more serious barriers in the future. So was born
HHReactor, which is the Producer-based implementation that lacks
lifetime manipulation (which I believe is cleaner),  e.g. `Using`.
However, it is much more compatible with the existing async
infrastructure in Hack, and so it will form the core, and HHRx will
become simply a wrapper of this.