Typechecks again, await testing
The treatment of Awaitable has been revamped to provide functionality
that is more natural to work with. `TotalAwaitable`, `Haltable` and
`ConditionWaitHandleWrapper` were all steps in the right direction to:

- Throw random Awaitables on the scheduler at any time from anywhere,
- Make an Awaitable whose value could be changed via interruption, a
  concept that is now exemplified by `Recyclable`, and
- Easily express "circular"-ish dependencies of notifier methods and
  ConditionWaitHandles (that wrap the lifetime of those notifiers) to
  generate Iterators from arbitrary time-events

respectively.

Their descendants are `Extendable[Lifetime]`, `Recyclable` and
EmitIterator & Subject respectively. `EmitIterator` and `Subject`
are particularly useful and interesting abstractions, because they
not only solve the headache of cyclic dependencies between
`ConditionWaitHandle` and notifier methods, but they also allow for
customizing the lifetime of the Iterator. The utility of this is
exemplified with memory-management for `Producer::window`, because
the `Producers` spawned must stop at either the next tick of the
signal, or the termination of the underlying `Producer`.

Interestingly, my first crack at the behavior of `Extendable` post-
termination has proven more useful than it is a hinderance -- I thought
at first that it would be overfit for the purpose of `Producer::
debounce`, but it found its way into `Subject` and is cooperating
with `Producer::flat_map`.

While some of the less straightforward methods in `Producer` were
refactored with the new classes mentioned above, `Producer::group_by`
didn't have much of an easy way through, _while_ maintaining the
performance of a `Map`. The most natural way to implement it with
`Subject` would require each eventual emitter to listen for their key
rather than the existing "I-call-you" model that is much more
performant.