Reverse yield and function roles in EmitIterator
Previously, yield pushed an Awaitable onto the scheduler, while a
function received values from the iterators and broadcasted them to
consumers. The motivation was mostly group_by, which could then
maintain a map of emitters, effectively reversing the control of the
emitting process.

There was a reason it was all so offputting in its convenience. It
was only a short hop from that implementation to Subject -- I was
effectively reversing control twice: from Iterator to "Subject"-ish
to Producer. And I hate Subject.

Turns out group_by can be implemented with wholesome iterators and
ConditionWaitHandle, so I could move back to the more idiomatic
role for yield: providing values to consumers. The emitters are
once again behind `private` walls, and Subject is dead.

"Back and forth. Forth and back." - Haywyre, 2013