Pop ready items from iterators at EmitIterator
AsyncIteratorWrapper is brought back from the dead, but is stripped of
halting. It is necessary because to pop ready items, I need to peek
inside the AsyncIterator, but once I call `next` once on the first
non-ready item, I can't call it again within the emit loop.

I should also mention the codebase typechecks again! Time for some
basic tests.