Fix two unrelated race conditions
The ready-wait core producer is very revealing when it comes to race
conditions. Case in point: the group_by->map->zip->flat_map flow that
was working with an asynchronous producer failed here since last
commit, and for three separate race conditions. Two are described in
the previous commit, but ever since migrating to Dependencies, some
undefined behavior that I must have been implicitly relying on
surfaced.

I've documented the debugging process privately (# 2017-04-09 10:32).
The race conditions fixed in this commit:

1. Some subproducers generated by `group_by` were failing early. Upon
closer inspection, the underlying Awaitable wasn't erring, so the
`ConditionWaitHandle` bell that wrapped it must have been completing
early. `group_by` ignores the classic `ConditionWaitHandle`
`InvalidArgumentException` because the timing of the last arc is
unknown to the scope resetting the handle. Originally, it was dependent
on a lifetime including `clone $this`, but while this raced pretty
obviously with the total iterator, I thought that it didn't matter
because I was thinking in Producer-lag mechanics. However, I was lazy
with the ad-hoc subjects so I didn't replicate `_pop_ready_wait_items`
behavior for the subjects. Then it did matter: the lag was filling but
the wait handle ended prematurely, ending the producer.

  The fix here is one possibility: it forces the dependency on the
total iterator. I'm still not sure if there is no race condition here,
because the producers produce independently, and if the bell-ringing
timing is weird, _maybe_ something could happen. I'm not totally sure
of the correctness of `_pop_ready_wait_handle` mechanics either, but
it's approaching the right impl.
2. Producer's total_awaitable wasn't depending on the scheduler. The
separate scheduler is a recent addition (last commit I believe) as
total_awaitable used to be Extendable, but even then only the original
async block was the only one coordinating the termination of the
Producer and was only depending on the original iterator. The success
of the original implementation was pure luck. Wrapping the original in
the Extendable, then adding the dependency in total_awaitable did the
trick.