Saving some progress trying out AsyncCondition
It's proving pretty hard to work with, because it's pretty inflexible
with the rules I want to put in place for it. Preferably, only the
child can notify a condition, but this means that the child trying to
make a Condition out of its own lifetime, or resettable ones -- really
all condition-related behavior, including the base-case, gets more
complicated. On dev-master, I'll continue working off the assumption
that `later` is guaranteed to block the context until control returns
back to the scheduler (either immediately for resumed wait handles, or
after the next sync arc for asyncfunctionwaithandles.