Tweak Producer for resource management
Producer has undergone yet another implementation change to manage its
resources when all of its consumers have disposed of their references.
Trying to follow the theme of minimizing custom methods for controlling
the producer, the producer listens for an unsubscription via __destruct,
or in the tree of producers, from a `detach` call. Future plan is to
implement `pause` with this mechanic too; it's nice that the option is
not far away.

Doing this required separating the emitters and the `next` call even
further than it already was. I was hesitant to step through the
iterators one element at a time and race their `next`s because usually
we want to return to the source of the values as quickly as possible --
usually a system call. Some extra flags ought to do the trick. I'm also
trying to make the bell much more optional so that I don't have to worry
about who/when is ringing it re: timing. If it's there, it'll be rung,
if not, then don't and proceed with the behavior (usually appending or
shifting the queues).

BaseProducer manages some of the flags and the destruct-detach behavior.
Why the separate class is a leftover from some of the process to get
here. Originally, I wanted AsyncGenerators only because they aren't
started even if they're functions (e.g. async () ==> { yield null; })
which gave me some nice guarantees with ConditionWH. However, since not
all async blocks are AsyncGenerators and I'm understanding why now (e.g.
`async { something_sync(); }` vs. `async { something_sync(); yield null;
}` do and don't eagerly call `something_sync` respectively), it was a
bit awkward with underlying functions that didn't yield, like the
`flat_map` implementation which only extends. So, whereby
`GeneratorProducer` was separate from `Producer`: the former wrapping
`AsyncGenerator` and the latter `AsyncIterator`, `AsyncIterator` as
before was more convenient to work with.