Fix cloning issues with flags and conditions
A few issues patched by this commit.

First, the appender was always putting the iterator on the schedule
and setting the `driver` with its lifetime, but this was assuming that
the appender could only be run after the first `next` on the Producer,
which goes back to when `Producer` wrapped `AsyncGenerator`s
exclusively. Instead, as discovered with a `collapse`d stream, this can
lead to mixed racecars, when the appended `driver`s are null, but the
initial `driver`s are set, plus the Producer has started. The solution:
wrap the `some_running` flag twice so that restarting after pausing
still doesn't race with the last arc of the awaitify loops from before
the pause, but the appender can have an up-to-date reference to the
current running state of the Producer.

Second, clones don't hold the same reference to the bell after it's
reset in one clone. Wrapping it fixes the problem: all of the clones
notify a common bell. Otherwise, one clone could run away with the
progress, and block all the other clones from advancing.

Third, small problem with the end condition of `_next`, where the buffer
isn't checked after the bell expires. I don't actually think this is
needed after making the bell common, since no element can be added to
the buffer unless the bell has been rung. If the bell expires without
ringing then, the buffer should be empty. This fix will probably go
away.