HTTPProducer alpha
HTTPProducer is able to listen and parse headers and stream bodies.
Occasionally, it will fail without warning from a "pure virual method
call". The cause is still unknown. It seems to arise when trying to
`await` a stream that has finished, but occasionally happens outside of
this case. More research is required.