Fix some stream reading logic: buffers & eof
Still feeling my way around stream_await and non-blocking streams. It
doesn't feel like checking the length of the `fread` fragment is the
best way to check if the buffer is empty. I've seen others advocating
for `feof`, but when the stream from the connection is made
non-blocking, for some reason it is never eof even when it's empty. My
guess is that, in that case, `feof` waits for the literal 0x04 (EOF), so
the only way to check if the buffer is empty is by trying to read and
checking the length (as it is now). Glad to be proved wrong about this.
This scheme is much more stable than the previous commit though.