Better define responsibilities for HTTP-server classes/functions
I have come to believe much of the HTTP server classes just transform
the system stream in almost a functional way. The iterator that produces
connections is the most obvious example, as it simply listens on the
stream for requests and tries to extract the headers, returning maybe a
bidirectional connection object if all the headers are sent through. The
bidirectional connection object can't be a function alone sadly because
it has to accept writes. Still, what distinguishes most of the classes
are how they interpret the incoming stream, so a Connection streaming a
body by content length will simply stop trying to iterate after the
content length has been reached, a WebSocket connection will parse the
data into frames and wait for the close opcode.

So, rather than composing, I hope more `Connection`s can extend the base
`Connection`, so that the purpose of a given connection is clearer (over
holding a plain Connection object somewhere and wrapping it.