Remove auto-clone for higher-order `Producer`s
While long ago, cloning was necessary to prevent race conditions between
competing consumers, a new system of producing has since replaced it,
which gives users the choice now to make warmer or cooler `Producer`s.
This automatic cloning forced the cool behavior (which was not very
cool) so it's been removed and instead a warning in the README serves to
point attention to this default and prevent any surprises.