Move exceptions into the buffer to propagate to consumers
Previously, I thought I could rely on the bell for error propagation,
but preparing for tests, it became obvious that, if the bell was rung
when an error sprung up in a child iterator, there would be no way to
propagate it aside from busy-waiting for the bell to be free again.

Moving it into the buffer also gives the consumers more freedom for
handling errors, since they can choose to catch and continue iterating
the remaining children (although that specific child must have stopped
from throwing the exception).