Add test dependencies and begin util classes
I'm opting to use HackUnit for the async support. I figure that, because
I can't test race conditions directly, the next best option is to make
it as likely for any that exist to manifest in tests. HackUnit default
behavior of parallelizing all async tests seems perfect for this, since
the racey parts are executed in as non-deterministic an order as
possible. While on the surface this goes against the golden rule to make
deterministic that which is not, since I can't mock a scheduler to
handle the `await` statements, this will at least expose the existence
of race conditions with the greatest likelihood.