React to window test
Many fantastic issues falling out of the `window` operator test!

1. The `Producer`s coming out of `window` need to be flattened to be
compared against 1D Vectors (avoiding 2D Vectors since the contents of
the inner Vectors aren't deterministic). However, while the flat_map
test passed, in this case the producers were just eager enough to cause
a race condition _within the constructor of `Producer`_! Setting
the `$racetrack` in the constructor via mapping the generator factories
raced against appending to the racetrack within `flat_map`. This was an
easy fix: initialize it by default to empty `Vector` to serve as a
target for these very eager append operations, and `concat` instead in
the constructor.
2. `window` looped infinitely when the signal was infinite because the
stop condition for the iterator waited only for the signal to end, not
the underlying producer. A "finished" flag served as a simple fix; I
needed essentially to limit the lifetime of the outer `window` result to
the minimum of the signal and underlying `Producer` lifetimes. If the
signal does finish first though, it will still continue producing on the
last window, which is the expected behavior.
3. I had realized that `window` was reproducing the values from the
beginning every time, and because it was cloning the base, unstarted
`Producer` in every window. Now, there is a direct race condition at the
"boundary" of two windows because if the signal ticks and the next
window picks up the next value before the previous, because they are
like two hot consumers, the second one will steal the value from the
first. This is rather unavoidable though without serious
ConditionWaitHandle acrobatics. I think that since the actual "tick" of
the signal is fuzzy, this is not very important anyways, but I'm open to
be proved wrong.