Guard against race condition on _append
If calling async functions, even synchronously, could indeed throw
control elsewhere, then before this commit `_append` had a race
condition, allowing the `Producer` to end before the appended iterator
had a chance to begin. This is because the `awaitify` call is a
synchronous invocation of an asynchronous function, and appending to the
racetrack is blocked on this call. Control could go to the last running
child iterator and end it, ending the `Producer` which has no knowledge
of this new iterator.