Repair destruction-detachment and `_produce`, `get_iterator_edge` race conditions
Previously, started `Producer`s weren't being destructed when they
became unreachable. This is because some of the closures that were
involved in iteration closed over `$this`, and the scheduler held
references to these closures, preventing them from being garbage
collected. There were a few interesting features of closures that this
fix needed to navigate around:

1. Closures close only over values that are mentioned within the closure
body. This, in hindsight, makes a lot of practical sense, and enabled
the fix because removing mentions of `$this` strategically allowed the
machinery to be separated from the `Producer` instance.
2. Methods that refer to `$this` seem to also create implicit references
to `$this` for the calling scope if the calling scope is a closure. This
meant that, although it seemed like `inst_meth` into a local variable
might remove the reference to `$this`, somehow the scheduler still clung
onto one, even if the local reference was disposed.

`awaitify` was the main target of exporting `$this`, but `_attach` had
to defer setting `driver` to `get_iterator_edge` (which is also able to
do a much cleaner job of it).

Fixes #4

---

Additionally, there is a race condition fix on `get_iterator_edge`.
`foreach` is sensitive to the collection changing, but knowing that
we're only appending, we can use a `for` loop instead.
`AwaitAllWaitHandle` somehow doesn't mind the collection changing while
it's being awaited, so rather than using `HH\Asio\v`, building an
`AwaitAllWaitHandle` out of a vector of the drivers desensitizes the
system to the race condition.