Implement debounce and address #1
A bit surprised by how compact the `debounce` implementation could end
up being. With reference to some thinking in #1, it hadn't crossed my
mind that I could shift the stream of values in time into a new stream
that I could consume and check for intervening values.

The test timing values are not totally arbitrary; their magnitude is so
large not for observability, but actually to suppress race conditions
(Macbook Pro Early 2015, 2.9GHz i5, ~1.8 load average). Times in the
10ms regime are so short that the consumer of the shifted iterator
within `debounce` was racing for control with the base iterator. In
every observed case, `debounce` produced too many items in this case,
but I haven't yet thought it through if this should always be expected.